# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareOS API — home care operating system foundation.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visits:
    get:
      operationId: listVisits
      tags: [visits]
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/VisitVerificationStatus"
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"
  /visits/clock-in:
    post:
      operationId: clockIn
      tags: [visits]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}/clock-out:
    post:
      operationId: clockOut
      tags: [visits]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
  /visits/{id}/verify:
    post:
      operationId: verifyVisit
      tags: [visits]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyVisitBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
...
components:
  schemas:
...
    VisitVerificationStatus:
      type: string
      enum: [PENDING, VERIFIED, EXCEPTION, REJECTED]
    ClockMethod:
      type: string
      enum: [GPS, TELEPHONY, FOB, MANUAL]
    Visit:
      type: object
      properties:
        id: { type: string }
        scheduleId: { type: string, nullable: true }
        caregiverId: { type: string }
        caregiverName: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        clockInTime: { type: string, format: date-time, nullable: true }
        clockInLat: { type: number, nullable: true }
        clockInLng: { type: number, nullable: true }
        clockInMethod:
          $ref: "#/components/schemas/ClockMethod"
        clockOutTime: { type: string, format: date-time, nullable: true }
        clockOutLat: { type: number, nullable: true }
        clockOutLng: { type: number, nullable: true }
        clockOutMethod:
          $ref: "#/components/schemas/ClockMethod"
        durationMinutes: { type: integer, nullable: true }
        tasksCompleted:
          type: array
          items: { type: string }
        caregiverNotes: { type: string, nullable: true }
        supervisorNotes: { type: string, nullable: true }
        verificationStatus:
          $ref: "#/components/schemas/VisitVerificationStatus"
        exceptionReason: { type: string, nullable: true }
        geoFenceMatch: { type: boolean }
      required:
        [
          id,
          caregiverId,
          caregiverName,
          clientId,
          clientName,
          clockInMethod,
          clockOutMethod,
          tasksCompleted,
          verificationStatus,
          geoFenceMatch,
        ]
    ClockInBody:
      type: object
      properties:
        scheduleId: { type: string }
        caregiverId: { type: string }
        clientId: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        method:
          $ref: "#/components/schemas/ClockMethod"
      required: [caregiverId, clientId, method]
    ClockOutBody:
      type: object
      properties:
        latitude: { type: number }
        longitude: { type: number }
        method:
          $ref: "#/components/schemas/ClockMethod"
        tasksCompleted:
          type: array
          items: { type: string }
        caregiverNotes: { type: string }
      required: [method]
    VerifyVisitBody:
      type: object
      properties:
        decision:
          type: string
          enum: [VERIFIED, REJECTED]
        supervisorNotes: { type: string }
      required: [decision]
...